Handle submit dine-in and clickandcollect
