+++ src/components/user/Checkout/CheckoutForm.js
@@ -11,20 +11,16 @@ import { useSelector } from "react-redux";
 import { formatPrice } from "../../Utils";
 import { useGuest } from "../../Context/guestInfos";
 import { useShoppingCart } from "../../Context/ShoppingCartContext";
-import { useParams } from "react-router-dom";
 
 const stripe = loadStripe(process.env.REACT_APP_STRIPE_PUBLIC_KEY, {
   betas: ["custom_checkout_beta_5"],
 });
 
 const CheckoutForm = ({ handleSubmit, isGuest }) => {
-  const { tableNumber } = useParams();
-
   const [savedInfo, setSavedInfo] = useState(false);
   const [clientSecret, setClientSecret] = useState(null);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
-  const [orderType, setOrderType] = useState("");
 
   const [cartData, setCartData] = useState([]);
 
@@ -44,7 +40,7 @@ const CheckoutForm = ({ handleSubmit, isGuest }) => {
 
   const user = useSelector((state) => state.userReducer);
 
-  const { cartItems, message, selectedDate, isClickAndCollect } =
+  const { message, selectedDate, calculateTotalPrice, items, orderType } =
     useShoppingCart();
 
   const stripeItems = useMemo(() => {
@@ -71,46 +67,9 @@ const CheckoutForm = ({ handleSubmit, isGuest }) => {
 
   const createCheckoutSession = useCallback(
     async (email) => {
-      const calculateTotalPrice = () => {
-        const total = cartItems.map((item) => {
-          const elPrice =
-            (formatPrice(item.price) +
-              (item.extraProteinPrice
-                ? parseFloat(item.extraProteinPrice)
-                : 0)) *
-            item.quantity;
-          return elPrice;
-        });
-
-        return total.reduce((acc, curr) => acc + curr).toFixed(2);
-      };
-
-      const items = cartItems.map((item) => {
-        const itemPrice =
-          (formatPrice(item.price) +
-            (item.extraProteinPrice ? parseFloat(item.extraProteinPrice) : 0)) *
-          item.quantity;
-
-        const meal = {
-          type: item.type,
-          name: item.name,
-          base: item.base,
-          proteins: item.proteins,
-          extraProtein: item.extraProtein,
-          extraProteinPrice: item.extraProteinPrice,
-          garnishes: item.garnishes,
-          toppings: item.toppings,
-          sauces: item.sauces,
-          quantity: item.quantity,
-          price: itemPrice,
-        };
-        return meal;
-      });
-
       const data = {
         userId: user._id ?? null,
         orderType,
-        ...(orderType === "dine-in" && { tableNumber }),
         items: items,
         specialInstructions: message,
         orderDate: selectedDate,
@@ -143,24 +102,14 @@ const CheckoutForm = ({ handleSubmit, isGuest }) => {
       message,
       orderType,
       selectedDate,
-      tableNumber,
       user._id,
       user.email,
       user.firstName,
       user.phone,
-      cartItems,
+      calculateTotalPrice,
+      items,
     ]
   );
-
-  useEffect(() => {
-    if (isClickAndCollect) {
-      setOrderType("clickandcollect");
-    } else if (tableNumber) {
-      setOrderType("dine-in");
-    } else {
-      setOrderType("clickandcollect");
-    }
-  }, [isClickAndCollect, tableNumber]);
 
   useEffect(() => {
     const storedCart = JSON.parse(sessionStorage.getItem("Cart")) || [];
